requirements.txt
streamlit
openai
langchain
langchain-community
langchain-openai
chromadb
python-dotenv
